feat: generate client entry for rsc
