+++ packages/ice/src/webpack/FlightClientEntryPlugin.ts
@@ -1,15 +1,90 @@
 import * as path from 'path';
 import { stringify } from 'querystring';
+import { asyncLib, acorn } from '@ice/bundles';
 import webpack from '@ice/bundles/compiled/webpack/index.js';
+import NullDependency from '@ice/bundles/compiled/webpack/NullDependency.js';
+import ModuleDependency from '@ice/bundles/compiled/webpack/ModuleDependency.js';
 import type { Compiler, Compilation } from 'webpack';
 
 const PLUGIN_NAME = 'FlightClientEntryPlugin';
 
+interface ClientReferenceSearchPath {
+  directory: string;
+  recursive?: boolean;
+  include: RegExp;
+  exclude?: RegExp;
+}
+
+interface Options {
+  clientReferences?: ClientReferenceSearchPath[];
+}
+
+class ClientReferenceDependency extends ModuleDependency {
+  userRequest: string;
+  request: string;
+
+  constructor(request: string) {
+    super(request);
+    this.request = request;
+  }
+  get type(): string {
+    return 'client-reference';
+  }
+}
+
 export class FlightClientEntryPlugin {
-  constructor() {
+  clientReferences: ClientReferenceSearchPath[];
+  clientFiles = new Set();
+
+  constructor(options: Options = {}) {
+    if (options.clientReferences) {
+      this.clientReferences = options.clientReferences;
+    } else {
+      this.clientReferences = [
+        {
+          directory: '.',
+          recursive: true,
+          include: /\.(js|ts|jsx|tsx)$/,
+          exclude: /types.ts|.d.ts|node_modules/,
+        },
+      ];
+    }
   }
 
   apply(compiler: Compiler) {
+    const _this = this;
+
+    compiler.hooks.thisCompilation.tap(PLUGIN_NAME, (compilation, { normalModuleFactory }) => {
+      // @ts-expect-error TODO: add types for ModuleDependency.
+      compilation.dependencyFactories.set(ClientReferenceDependency, normalModuleFactory);
+      // @ts-expect-error TODO: add types for ModuleDependency.
+      compilation.dependencyTemplates.set(ClientReferenceDependency, new NullDependency.Template());
+    });
+
+    compiler.hooks.beforeCompile.tapAsync(PLUGIN_NAME, ({ contextModuleFactory }, callback) => {
+      const contextResolver = compiler.resolverFactory.get('context', {});
+      const normalResolver = compiler.resolverFactory.get('normal');
+
+      _this.resolveClientFiles(
+        compiler.context,
+        contextResolver,
+        normalResolver,
+        compiler.inputFileSystem,
+        contextModuleFactory,
+        (err, resolvedClientRefs) => {
+          if (err) {
+            callback(err);
+            return;
+          }
+          resolvedClientRefs.forEach(dep => {
+            this.clientFiles.add(dep.request);
+          });
+
+          callback();
+        },
+      );
+    });
+
     compiler.hooks.finishMake.tapPromise(PLUGIN_NAME, (compilation) =>
       this.createClientEntries(compiler, compilation),
     );
@@ -40,8 +115,6 @@ export class FlightClientEntryPlugin {
         });
 
         if (clientComponentImports.length || CSSImports.length) {
-          console.log(name, entryRequest, clientComponentImports, CSSImports);
-
           const injected = this.injectClientEntry({
             compiler,
             compilation,
@@ -73,7 +146,6 @@ export class FlightClientEntryPlugin {
     return new Promise<void>((resolve, reject) => {
       compilation.addEntry(compiler.context, clientComponentEntryDep, { name, dependOn: ['main'] }, (err) => {
         if (err) {
-          console.error(err);
           reject(err);
         }
 
@@ -103,7 +175,7 @@ export class FlightClientEntryPlugin {
         CSSImports.push(modRequest);
       }
 
-      if (isClientComponentEntryModule(mod)) {
+      if (this.isClientComponentEntryModule(mod)) {
         clientComponentImports.push(modRequest);
         return;
       }
@@ -120,6 +192,118 @@ export class FlightClientEntryPlugin {
       clientComponentImports,
       CSSImports,
     };
+  }
+
+  isClientComponentEntryModule(mod) {
+    if (this.clientFiles.has(mod.resource)) {
+      return true;
+    }
+
+    return false;
+  }
+
+  resolveClientFiles(
+    context: string,
+    contenxtResolver: ReturnType<Compiler['resolverFactory']['get']>,
+    normalResolver: ReturnType<Compiler['resolverFactory']['get']>,
+    fs: Compilation['inputFileSystem'],
+    contextModuleFactory: Compilation['params']['contextModuleFactory'],
+    callback: (err: Error | null, files?: ClientReferenceDependency[]) => void,
+  ) {
+    function hasUseClientDirective(source: string): boolean {
+      if (source.indexOf('use client') === -1) {
+        return false;
+      }
+      let body;
+      try {
+        // TODO: check client directive by comment injected by swc plugin.
+        body = acorn.parse(source, {
+          ecmaVersion: '2024',
+          sourceType: 'module',
+        }).body;
+      } catch (x) {
+        return false;
+      }
+      for (let i = 0; i < body.length; i++) {
+        const node = body[i];
+        if (node.type !== 'ExpressionStatement' || !node.directive) {
+          break;
+        }
+        if (node.directive === 'use client') {
+          return true;
+        }
+      }
+      return false;
+    }
+    asyncLib.map(this.clientReferences, (
+      clientReference: ClientReferenceSearchPath,
+      cb: (err: null | Error, result?: ClientReferenceDependency[]) => void,
+    ) => {
+      contenxtResolver.resolve({}, context, clientReference.directory, {}, (err, resolvedDirectory) => {
+        if (err) return cb(err);
+        const options = {
+          resource: resolvedDirectory,
+          resourceQuery: '',
+          recursive:
+          clientReference.recursive === undefined
+            ? true
+            : clientReference.recursive,
+          regExp: clientReference.include,
+          include: undefined,
+          exclude: clientReference.exclude,
+        };
+        // @ts-expect-error TODO: add types for resolveDependencies options.
+        contextModuleFactory.resolveDependencies(fs, options, (err, dependencies) => {
+          if (err) return cb(err);
+          const clientRefDeps = dependencies.map(dep => {
+            // Use userRequest instead of request. request always end with undefined which is wrong.
+            const request = path.join(resolvedDirectory as string, dep.userRequest);
+            const clientRefDep = new ClientReferenceDependency(request);
+            clientRefDep.userRequest = dep.userRequest;
+            return clientRefDep;
+          });
+          asyncLib.filter(
+            clientRefDeps,
+            (dep: ClientReferenceDependency, filterCb: (err: null | Error, truthValue: boolean) => void,
+          ) => {
+            normalResolver.resolve(
+              {},
+              context,
+              dep.request,
+              {},
+              (err: null | Error, resolvedPath: any) => {
+                if (err || typeof resolvedPath !== 'string') {
+                  return filterCb(null, false);
+                }
+
+                fs.readFile(
+                  resolvedPath,
+                  'utf-8',
+                  // @ts-expect-error
+                  (err: null | Error, content: string) => {
+                    if (err || typeof content !== 'string') {
+                      return filterCb(null, false);
+                    }
+                    const useClient = hasUseClientDirective(content);
+                    filterCb(null, useClient);
+                  },
+                );
+              },
+            );
+          }, cb);
+        });
+      });
+    }, (
+      err: null | Error,
+      result: ClientReferenceDependency[],
+    ) => {
+      if (err) return callback(err);
+      const flat: ClientReferenceDependency[] = [];
+      for (let i = 0; i < result.length; i++) {
+        flat.push.apply(flat, result[i]);
+      }
+      callback(null, flat);
+    });
   }
 }
